[RFR] Extract advertising, qwr, and connection params initialization
