+++ main.c
@@ -32,35 +32,24 @@
 
 #include "modules/ble/service/service.h"
 #include "modules/ble/service/characteristic.h"
+#include "modules/ble/gap/advertising.h"
+#include "modules/ble/ble_helpers.h"
 
 #define DEVICE_NAME "SomeLongName"              /**< Extended name of a device. */
 #define MANUFACTURER_NAME "NordicSemiconductor" /**< Manufacturer. Will be passed to Device Information Service. */
-#define APP_ADV_INTERVAL 300                    /**< The advertising interval (in units of 0.625 ms. This value corresponds to 187.5 ms). */
-
-#define APP_ADV_DURATION 18000  /**< The advertising duration (180 seconds) in units of 10 milliseconds. */
-#define APP_BLE_OBSERVER_PRIO 3 /**< Application's BLE observer priority. You shouldn't need to modify this value. */
-#define APP_BLE_CONN_CFG_TAG 1  /**< A tag identifying the SoftDevice BLE configuration. */
 
 #define MIN_CONN_INTERVAL MSEC_TO_UNITS(100, UNIT_1_25_MS) /**< Minimum acceptable connection interval (0.1 seconds). */
 #define MAX_CONN_INTERVAL MSEC_TO_UNITS(200, UNIT_1_25_MS) /**< Maximum acceptable connection interval (0.2 second). */
 #define SLAVE_LATENCY 0                                    /**< Slave latency. */
 #define CONN_SUP_TIMEOUT MSEC_TO_UNITS(4000, UNIT_10_MS)   /**< Connection supervisory timeout (4 seconds). */
 
-#define FIRST_CONN_PARAMS_UPDATE_DELAY APP_TIMER_TICKS(5000) /**< Time from initiating event (connect or start of notification) to first time sd_ble_gap_conn_param_update is called (5 seconds). */
-#define NEXT_CONN_PARAMS_UPDATE_DELAY APP_TIMER_TICKS(30000) /**< Time between each call to sd_ble_gap_conn_param_update after the first call (30 seconds). */
-#define MAX_CONN_PARAMS_UPDATE_COUNT 3                       /**< Number of attempts before giving up the connection parameter negotiation. */
-
 #define DEAD_BEEF 0xDEADBEEF /**< Value used as error code on stack dump, can be used to identify stack location on stack unwind. */
 
-#define ESTC_SOME_CHARACTERISTIC_UUID16  0x05b9
+#define ESTC_SOME_CHARACTERISTIC_UUID16 0x05b9
 
-static uint8_t some_characteristic_desc[] = { 0x73, 0x6f, 0x6d, 0x65, 0x20, 0x64, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x20, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67};
+static uint8_t some_characteristic_desc[] = {0x73, 0x6f, 0x6d, 0x65, 0x20, 0x64, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x20, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67};
 
 NRF_BLE_GATT_DEF(m_gatt);           /**< GATT module instance. */
-NRF_BLE_QWR_DEF(m_qwr);             /**< Context for the Queued Write module.*/
-BLE_ADVERTISING_DEF(m_advertising); /**< Advertising module instance. */
-
-static uint16_t m_conn_handle = BLE_CONN_HANDLE_INVALID; /**< Handle of the current connection. */
 
 static ble_uuid_t m_adv_uuids[] = /**< Universally unique service identifiers. */
 {
@@ -72,12 +61,10 @@ static ble_uuid_t m_adv_uuids[] = /**< Universally unique service identifiers. *
 
 static estc_ble_service_t m_estc_service =
 {
-    .service_handle = 7198U,
+    .service_handle = 0U,
 };
 
 static uint32_t some_characteristic_value = 32U;
-
-static void advertising_start(void);
 
 /**@brief Callback function for asserts in the SoftDevice.
  *
@@ -146,44 +133,27 @@ static void gatt_init(void)
     APP_ERROR_CHECK(err_code);
 }
 
-/**@brief Function for handling Queued Write Module errors.
- *
- * @details A pointer to this function will be passed to each service which may need to inform the
- *          application about an error.
- *
- * @param[in]   nrf_error   Error code containing information about what went wrong.
- */
-static void nrf_qwr_error_handler(uint32_t nrf_error)
-{
-    APP_ERROR_HANDLER(nrf_error);
-}
-
 /**@brief Function for initializing services that will be used by the application.
  */
 static void services_init(void)
 {
     ret_code_t err_code;
-    nrf_ble_qwr_init_t qwr_init = {0};
 
-    // Initialize Queued Write Module.
-    qwr_init.error_handler = nrf_qwr_error_handler;
-
-    err_code = nrf_ble_qwr_init(&m_qwr, &qwr_init);
-    APP_ERROR_CHECK(err_code);
+    estc_ble_qwr_init();
 
     err_code = estc_ble_service_init(&m_estc_service);
     APP_ERROR_CHECK(err_code);
 
-    estc_ble_service_characteristic_config_t dummy_config = 
+    estc_ble_service_characteristic_config_t dummy_config =
     {
-        .characteristic_uuid      = ESTC_SOME_CHARACTERISTIC_UUID16,
-        .description_string       = some_characteristic_desc,
-        .description_size         = sizeof(some_characteristic_desc),
-        .permissions              = ESTC_BLE_SERVICE_CHARACTERISTIC_ALLOW_READ | ESTC_BLE_SERVICE_CHARACTERISTIC_ALLOW_WRITE,
-        .value                    = (uint8_t * )(&some_characteristic_value),
-        .value_size               = sizeof(some_characteristic_value),
-        .value_type               = ESTC_BLE_SERVICE_CHARACTERISTIC_TYPE_UINT32,
-        .value_size_max           = sizeof(some_characteristic_value),
+        .characteristic_uuid = ESTC_SOME_CHARACTERISTIC_UUID16,
+        .description_string = some_characteristic_desc,
+        .description_size = sizeof(some_characteristic_desc),
+        .permissions = ESTC_BLE_SERVICE_CHARACTERISTIC_ALLOW_READ | ESTC_BLE_SERVICE_CHARACTERISTIC_ALLOW_WRITE,
+        .value = (uint8_t *)(&some_characteristic_value),
+        .value_size = sizeof(some_characteristic_value),
+        .value_type = ESTC_BLE_SERVICE_CHARACTERISTIC_TYPE_UINT32,
+        .value_size_max = sizeof(some_characteristic_value),
         .characteristic_uuid_type = BLE_UUID_TYPE_BLE,
     };
 
@@ -193,149 +163,10 @@ static void services_init(void)
     APP_ERROR_CHECK(err_code);
 }
 
-/**@brief Function for handling the Connection Parameters Module.
- *
- * @details This function will be called for all events in the Connection Parameters Module which
- *          are passed to the application.
- *          @note All this function does is to disconnect. This could have been done by simply
- *                setting the disconnect_on_fail config parameter, but instead we use the event
- *                handler mechanism to demonstrate its use.
- *
- * @param[in] p_evt  Event received from the Connection Parameters Module.
- */
-static void on_conn_params_evt(ble_conn_params_evt_t *p_evt)
-{
-    ret_code_t err_code;
-
-    if (p_evt->evt_type == BLE_CONN_PARAMS_EVT_FAILED)
-    {
-        err_code = sd_ble_gap_disconnect(m_conn_handle, BLE_HCI_CONN_INTERVAL_UNACCEPTABLE);
-        APP_ERROR_CHECK(err_code);
-    }
-}
-
-/**@brief Function for handling a Connection Parameters error.
- *
- * @param[in] nrf_error  Error code containing information about what went wrong.
- */
-static void conn_params_error_handler(uint32_t nrf_error)
-{
-    APP_ERROR_HANDLER(nrf_error);
-}
-
-/**@brief Function for initializing the Connection Parameters module.
- */
-static void conn_params_init(void)
-{
-    ret_code_t err_code;
-    ble_conn_params_init_t cp_init;
-
-    memset(&cp_init, 0, sizeof(cp_init));
-
-    cp_init.p_conn_params = NULL;
-    cp_init.first_conn_params_update_delay = FIRST_CONN_PARAMS_UPDATE_DELAY;
-    cp_init.next_conn_params_update_delay = NEXT_CONN_PARAMS_UPDATE_DELAY;
-    cp_init.max_conn_params_update_count = MAX_CONN_PARAMS_UPDATE_COUNT;
-    cp_init.start_on_notify_cccd_handle = BLE_GATT_HANDLE_INVALID;
-    cp_init.disconnect_on_fail = false;
-    cp_init.evt_handler = on_conn_params_evt;
-    cp_init.error_handler = conn_params_error_handler;
-
-    err_code = ble_conn_params_init(&cp_init);
-    APP_ERROR_CHECK(err_code);
-}
-
 /**@brief Function for starting timers.
  */
 static void application_timers_start(void)
 {
-}
-
-/**@brief Function for handling advertising events.
- *
- * @details This function will be called for advertising events which are passed to the application.
- *
- * @param[in] ble_adv_evt  Advertising event.
- */
-static void on_adv_evt(ble_adv_evt_t ble_adv_evt)
-{
-    ret_code_t err_code;
-
-    switch (ble_adv_evt)
-    {
-    case BLE_ADV_EVT_FAST:
-        NRF_LOG_INFO("Fast advertising.");
-        err_code = bsp_indication_set(BSP_INDICATE_ADVERTISING);
-        APP_ERROR_CHECK(err_code);
-        break;
-
-    default:
-        break;
-    }
-}
-
-/**@brief Function for handling BLE events.
- *
- * @param[in]   p_ble_evt   Bluetooth stack event.
- * @param[in]   p_context   Unused.
- */
-static void ble_evt_handler(ble_evt_t const *p_ble_evt, void *p_context)
-{
-    ret_code_t err_code = NRF_SUCCESS;
-
-    switch (p_ble_evt->header.evt_id)
-    {
-    case BLE_GAP_EVT_DISCONNECTED:
-        NRF_LOG_INFO("Disconnected.");
-        // LED indication will be changed when advertising starts.
-        break;
-
-    case BLE_GAP_EVT_CONNECTED:
-        NRF_LOG_INFO("Connected.");
-        err_code = bsp_indication_set(BSP_INDICATE_CONNECTED);
-        APP_ERROR_CHECK(err_code);
-        m_conn_handle = p_ble_evt->evt.gap_evt.conn_handle;
-        err_code = nrf_ble_qwr_conn_handle_assign(&m_qwr, m_conn_handle);
-        APP_ERROR_CHECK(err_code);
-        break;
-
-    case BLE_GAP_EVT_PHY_UPDATE_REQUEST:
-    {
-        NRF_LOG_DEBUG("PHY update request.");
-        ble_gap_phys_t const phys =
-            {
-                .rx_phys = BLE_GAP_PHY_AUTO,
-                .tx_phys = BLE_GAP_PHY_AUTO,
-            };
-        err_code = sd_ble_gap_phy_update(p_ble_evt->evt.gap_evt.conn_handle, &phys);
-        APP_ERROR_CHECK(err_code);
-    }
-    break;
-
-    case BLE_GATTC_EVT_TIMEOUT:
-        // Disconnect on GATT Client timeout event.
-        NRF_LOG_DEBUG("GATT Client Timeout.");
-        err_code = sd_ble_gap_disconnect(p_ble_evt->evt.gattc_evt.conn_handle,
-                                         BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
-        APP_ERROR_CHECK(err_code);
-        break;
-
-    case BLE_GATTS_EVT_TIMEOUT:
-        // Disconnect on GATT Server timeout event.
-        NRF_LOG_DEBUG("GATT Server Timeout.");
-        err_code = sd_ble_gap_disconnect(p_ble_evt->evt.gatts_evt.conn_handle,
-                                         BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
-        APP_ERROR_CHECK(err_code);
-        break;
-
-    case BLE_GATTS_EVT_WRITE:
-        NRF_LOG_DEBUG("GATT Server Write.");
-        
-
-    default:
-        // No implementation needed.
-        break;
-    }
 }
 
 /**@brief Function for initializing the BLE stack.
@@ -360,60 +191,7 @@ static void ble_stack_init(void)
     APP_ERROR_CHECK(err_code);
 
     // Register a handler for BLE events.
-    NRF_SDH_BLE_OBSERVER(m_ble_observer, APP_BLE_OBSERVER_PRIO, ble_evt_handler, NULL);
-}
-
-/**@brief Function for handling events from the BSP module.
- *
- * @param[in]   event   Event generated when button is pressed.
- */
-static void bsp_event_handler(bsp_event_t event)
-{
-    ret_code_t err_code;
-
-    switch (event)
-    {
-    case BSP_EVENT_DISCONNECT:
-        err_code = sd_ble_gap_disconnect(m_conn_handle,
-                                         BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
-        if (err_code != NRF_ERROR_INVALID_STATE)
-        {
-            APP_ERROR_CHECK(err_code);
-        }
-        break; // BSP_EVENT_DISCONNECT
-    default:
-        break;
-    }
-}
-
-/**@brief Function for initializing the Advertising functionality.
- */
-static void advertising_init(void)
-{
-    ret_code_t err_code;
-    ble_advertising_init_t init;
-
-    memset(&init, 0, sizeof(init));
-
-    init.advdata.name_type = BLE_ADVDATA_NO_NAME;
-    init.advdata.flags = BLE_GAP_ADV_FLAGS_LE_ONLY_GENERAL_DISC_MODE;
-
-    // init.advdata.p_tx_power_level       = &tx_power_level;
-
-    init.srdata.name_type = BLE_ADVDATA_FULL_NAME;
-    init.srdata.uuids_complete.uuid_cnt = sizeof(m_adv_uuids) / sizeof(m_adv_uuids[0]);
-    init.srdata.uuids_complete.p_uuids = m_adv_uuids;
-
-    init.config.ble_adv_fast_enabled = true;
-    init.config.ble_adv_fast_interval = APP_ADV_INTERVAL;
-    init.config.ble_adv_fast_timeout = APP_ADV_DURATION;
-
-    init.evt_handler = on_adv_evt;
-
-    err_code = ble_advertising_init(&m_advertising, &init);
-    APP_ERROR_CHECK(err_code);
-
-    ble_advertising_conn_cfg_tag_set(&m_advertising, APP_BLE_CONN_CFG_TAG);
+    NRF_SDH_BLE_OBSERVER(m_ble_observer, APP_BLE_OBSERVER_PRIO, estc_ble_default_event_handler, NULL);
 }
 
 /**@brief Function for initializing buttons and leds.
@@ -424,7 +202,7 @@ static void buttons_leds_init(void)
 {
     ret_code_t err_code;
 
-    err_code = bsp_init(BSP_INIT_LEDS | BSP_INIT_BUTTONS, bsp_event_handler);
+    err_code = bsp_init(BSP_INIT_LEDS | BSP_INIT_BUTTONS, estc_bsp_default_event_handler);
     APP_ERROR_CHECK(err_code);
 
     err_code = bsp_btn_ble_init(NULL, NULL);
@@ -463,14 +241,6 @@ static void idle_state_handle(void)
     LOG_BACKEND_USB_PROCESS();
 }
 
-/**@brief Function for starting advertising.
- */
-static void advertising_start(void)
-{
-    ret_code_t err_code = ble_advertising_start(&m_advertising, BLE_ADV_MODE_FAST);
-    APP_ERROR_CHECK(err_code);
-}
-
 /**@brief Function for application main entry.
  */
 int main(void)
@@ -483,15 +253,15 @@ int main(void)
     ble_stack_init();
     gap_params_init();
     gatt_init();
-    advertising_init();
+    estc_ble_gap_advertising_init(m_adv_uuids, sizeof(m_adv_uuids) / sizeof(m_adv_uuids[0]));
     services_init();
-    conn_params_init();
+    estc_ble_conn_params_init();
 
     // Start execution.
     NRF_LOG_INFO("ESTC advertising example started.");
     application_timers_start();
 
-    advertising_start();
+    estc_ble_gap_advertising_start();
 
     // Enter main loop.
     for (;;)
