+++ modules/ble/ble_helpers.c
@@ -1,0 +1,166 @@
+#include <app_error.h>
+#include <nrf_ble_qwr.h>
+#include <bsp_btn_ble.h>
+#include <nrf_sdh_ble.h>
+#include <nrf_log.h>
+#include <app_timer.h>
+
+#include "ble_helpers.h"
+
+NRF_BLE_QWR_DEF(m_qwr); /**< Context for the Queued Write module.*/
+
+static uint16_t m_conn_handle = BLE_CONN_HANDLE_INVALID; /**< Handle of the current connection. */
+
+/**
+ * @brief Function for handling Queued Write Module errors.
+ *
+ * @details A pointer to this function will be passed to each service which may need to inform the
+ *          application about an error.
+ *
+ * @param[in]   nrf_error   Error code containing information about what went wrong.
+ */
+static void nrf_qwr_error_handler(uint32_t nrf_error)
+{
+	APP_ERROR_HANDLER(nrf_error);
+}
+
+/**
+ * @brief Function for handling the Connection Parameters Module.
+ *
+ * @details This function will be called for all events in the Connection Parameters Module which
+ *          are passed to the application.
+ *          @note All this function does is to disconnect. This could have been done by simply
+ *                setting the disconnect_on_fail config parameter, but instead we use the event
+ *                handler mechanism to demonstrate its use.
+ *
+ * @param[in] p_evt  Event received from the Connection Parameters Module.
+ */
+static void estc_ble_on_conn_params_evt_handler(ble_conn_params_evt_t *p_evt)
+{
+	ret_code_t err_code;
+
+	if (p_evt->evt_type == BLE_CONN_PARAMS_EVT_FAILED)
+	{
+		err_code = sd_ble_gap_disconnect(m_conn_handle, BLE_HCI_CONN_INTERVAL_UNACCEPTABLE);
+		APP_ERROR_CHECK(err_code);
+	}
+}
+
+/**
+ * @brief Function for handling a Connection Parameters error.
+ *
+ * @param[in] nrf_error  Error code containing information about what went wrong.
+ */
+static void estc_conn_params_error_handler(uint32_t nrf_error)
+{
+	APP_ERROR_HANDLER(nrf_error);
+}
+
+void estc_ble_qwr_init(void)
+{
+	ret_code_t err_code = NRF_SUCCESS;
+	nrf_ble_qwr_init_t qwr_init = {0};
+
+	// Initialize Queued Write Module.
+	qwr_init.error_handler = nrf_qwr_error_handler;
+
+	err_code = nrf_ble_qwr_init(&m_qwr, &qwr_init);
+	APP_ERROR_CHECK(err_code);
+}
+
+void estc_ble_default_event_handler(ble_evt_t const *p_ble_evt, void *p_context)
+{
+	ret_code_t err_code = NRF_SUCCESS;
+
+	switch (p_ble_evt->header.evt_id)
+	{
+	case BLE_GAP_EVT_DISCONNECTED:
+		NRF_LOG_INFO("Disconnected.");
+		// LED indication will be changed when advertising starts.
+		break;
+
+	case BLE_GAP_EVT_CONNECTED:
+		NRF_LOG_INFO("Connected.");
+		err_code = bsp_indication_set(BSP_INDICATE_CONNECTED);
+		APP_ERROR_CHECK(err_code);
+		m_conn_handle = p_ble_evt->evt.gap_evt.conn_handle;
+		err_code = nrf_ble_qwr_conn_handle_assign(&m_qwr, p_ble_evt->evt.gap_evt.conn_handle);
+		APP_ERROR_CHECK(err_code);
+		break;
+
+	case BLE_GAP_EVT_PHY_UPDATE_REQUEST:
+	{
+		NRF_LOG_DEBUG("PHY update request.");
+		ble_gap_phys_t const phys =
+			{
+				.rx_phys = BLE_GAP_PHY_AUTO,
+				.tx_phys = BLE_GAP_PHY_AUTO,
+			};
+		err_code = sd_ble_gap_phy_update(p_ble_evt->evt.gap_evt.conn_handle, &phys);
+		APP_ERROR_CHECK(err_code);
+	}
+	break;
+
+	case BLE_GATTC_EVT_TIMEOUT:
+		// Disconnect on GATT Client timeout event.
+		NRF_LOG_DEBUG("GATT Client Timeout.");
+		err_code = sd_ble_gap_disconnect(p_ble_evt->evt.gattc_evt.conn_handle,
+										 BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
+		APP_ERROR_CHECK(err_code);
+		break;
+
+	case BLE_GATTS_EVT_TIMEOUT:
+		// Disconnect on GATT Server timeout event.
+		NRF_LOG_DEBUG("GATT Server Timeout.");
+		err_code = sd_ble_gap_disconnect(p_ble_evt->evt.gatts_evt.conn_handle,
+										 BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
+		APP_ERROR_CHECK(err_code);
+		break;
+
+	case BLE_GATTS_EVT_WRITE:
+		NRF_LOG_DEBUG("GATT Server Write.");
+
+	default:
+		// No implementation needed.
+		break;
+	}
+}
+
+void estc_bsp_default_event_handler(bsp_event_t event)
+{
+	ret_code_t err_code;
+
+	switch (event)
+	{
+	case BSP_EVENT_DISCONNECT:
+		err_code = sd_ble_gap_disconnect(m_conn_handle,
+										 BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
+		if (err_code != NRF_ERROR_INVALID_STATE)
+		{
+			APP_ERROR_CHECK(err_code);
+		}
+		break; // BSP_EVENT_DISCONNECT
+	default:
+		break;
+	}
+}
+
+void estc_ble_conn_params_init(void)
+{
+	ret_code_t err_code;
+	ble_conn_params_init_t cp_init;
+
+	memset(&cp_init, 0, sizeof(cp_init));
+
+	cp_init.p_conn_params = NULL;
+	cp_init.first_conn_params_update_delay = FIRST_CONN_PARAMS_UPDATE_DELAY;
+	cp_init.next_conn_params_update_delay = NEXT_CONN_PARAMS_UPDATE_DELAY;
+	cp_init.max_conn_params_update_count = MAX_CONN_PARAMS_UPDATE_COUNT;
+	cp_init.start_on_notify_cccd_handle = BLE_GATT_HANDLE_INVALID;
+	cp_init.disconnect_on_fail = false;
+	cp_init.evt_handler = estc_ble_on_conn_params_evt_handler;
+	cp_init.error_handler = estc_conn_params_error_handler;
+
+	err_code = ble_conn_params_init(&cp_init);
+	APP_ERROR_CHECK(err_code);
+}
